+++ src/fcn/weather/getArchiveWeatherData.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Acuparse - AcuRite®‎ Access/smartHUB and IP Camera Data Processing, Display, and Upload.
- * @copyright Copyright (C) 2015-2019 Maxwell Power
+ * Acuparse - AcuRite Access/smartHUB and IP Camera Data Processing, Display, and Upload.
+ * @copyright Copyright (C) 2015-2020 Maxwell Power
  * @author Maxwell Power <max@acuparse.com>
  * @link http://www.acuparse.com
  * @license AGPL-3.0+
@@ -523,7 +523,7 @@ class getArchiveWeatherData
         $this->rainfall_MM_most_ever = (float)round($result['dailyrainin'] * 25.4, 2); // Millimeters
     }
 
-    //Private Functions
+    // Private Functions
 
     // Calculate human readable wind direction from a range of values:
     private function windDirection($windDEG)