+++ src/pub/admin/index.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Acuparse - AcuRite®‎ Access/smartHUB and IP Camera Data Processing, Display, and Upload.
- * @copyright Copyright (C) 2015-2019 Maxwell Power
+ * Acuparse - AcuRite Access/smartHUB and IP Camera Data Processing, Display, and Upload.
+ * @copyright Copyright (C) 2015-2020 Maxwell Power
  * @author Maxwell Power <max@acuparse.com>
  * @link http://www.acuparse.com
  * @license AGPL-3.0+
@@ -76,6 +76,11 @@ if (isset($_SESSION['authenticated']) && $_SESSION['authenticated'] === true && 
                 <button type="button" id="system-settings" class="btn btn-outline-secondary btn-block"
                         onclick="location.href = '/admin/settings'"><i class="fas fa-cogs"
                                                                        aria-hidden="true"></i> System Settings
+                </button>
+                <h4 class="margin-top-05">Sensors:</h4>
+                <button type="button" id="system-settings" class="btn btn-outline-secondary btn-block"
+                        onclick="location.href = '/admin/status'"><i class="fas fa-question-circle"
+                                                                       aria-hidden="true"></i> Sensor Status
                 </button>
                 <h4 class="margin-top-05">Access Tools:</h4>
                 <button type="button" id="access-server" class="btn btn-outline-secondary btn-block margin-top-10"