+++ CHANGELOG.md
@@ -7,11 +7,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [[3.0.0]](https://www.acuparse.com/releases/v3-0-0/) - BETA
+
+### Added
+
+- Support all Atlas readings (UV, Lightning, Light).
+
+### Fixed
+
+- MySQL error handling.
+
+### Changed
+
+- Bootstrap to v4.4.1.
+- FontAwesome to v5.12.1.
+- instant.page to v3.0.0.
+- Update Copyright.
+
 ## [[2.10.0]](https://www.acuparse.com/releases/v2-10-0/) - 2019-10-06
 
 ### Added
 
-- Support for Lightning Tower.
+- Basic support for Lightning Tower.
 - Support sending unknown sensors to MyAcuRite.
 
 ### Changed
@@ -43,10 +60,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Lightbox to v2.11.1
 - InstantPage to v2.0.0
 - Mailing list to Google Apps
-
-### Added
-
-- Gitter Badge
 
 ## [[2.9.1]](https://www.acuparse.com/releases/v2-9-1/) - 2019-06-28